feat: add option for decimal prices

+++ src/main/java/com/spygstudios/chestshop/shop/AmountHandler.java
@@ -36,7 +36,7 @@ public class AmountHandler {
         pendingAmount.add(this);
     }
 
-    public void create(int amount) {
+    public void create(double amount) {
         if (type.equals(GuiAction.SET_ITEM_PRICE)) {
             shop.setPrice(amount);
         }
